feat: connect year graph API
